mainHack.ns: only target another server after a hack
Early in the game, it happened too often that the script selected some server
xxxx for weaking/growing. But because of that, the script gained hack experience
for the player. And on the next loop, it therefore selected another server as
target.

The effect was that no hacks were done for a longer time, and this money needed
for Tor router or programs didn't trickle in.

This patch changes the behavior of the script so that it only selects a new
bestTarget if a hack finished.